+++ objc2-foundation/build.rs
@@ -1,9 +1,0 @@
-use std::env;
-
-fn main() {
-    // The script doesn't depend on our code
-    println!("cargo:rerun-if-changed=build.rs");
-
-    let runtime = env::var("DEP_OBJC_0_2_RUNTIME").unwrap();
-    println!("cargo:rustc-cfg={}", runtime);
-}