+++ objc2/src/rc/id.rs
@@ -308,7 +308,7 @@ impl<T: Message, O: Ownership> Id<T, O> {
 
         // Only worth doing on the Apple runtime.
         // Not supported on TARGET_OS_WIN32.
-        #[cfg(all(apple, not(target_os = "windows")))]
+        #[cfg(all(feature = "apple", not(target_os = "windows")))]
         {
             // Supported since macOS 10.7.
             #[cfg(target_arch = "x86_64")]
@@ -352,7 +352,7 @@ impl<T: Message, O: Ownership> Id<T, O> {
         // less likely to occur.
         //
         // This is brittle! We should find a better solution!
-        #[cfg(all(apple, not(target_os = "windows"), target_arch = "x86_64"))]
+        #[cfg(all(feature = "apple", not(target_os = "windows"), target_arch = "x86_64"))]
         {
             // SAFETY: Similar to above.
             unsafe { core::arch::asm!("nop", options(nomem, preserves_flags, nostack)) };